chore: type import 관련 수정
